MetricsAgentTests/CpuMetricsControllerUnitTest.cs
using MetricsAgent.Controllers;
using MetricsAgent.DAL;
using MetricsAgent.Models;
using Microsoft.AspNetCore.Mvc;
using Moq;
using System;
using Xunit;

namespace MetricsAgentTests
{
    public class CpuMetricsControllerUnitTest
    {
        private CpuMetricsController _controller;
        private Mock<ICpuMetricsRepository> _mock;

        public CpuMetricsControllerUnitTest()
        {
            _mock = new Mock<ICpuMetricsRepository>();
            _controller = new CpuMetricsController(_mock.Object);
        }

        [Fact]
        public void CallCreateTest()
        {
            _mock.Setup(repository => repository.Create(It.IsAny<BaseMetricValue>())).Verifiable();

            var result = _controller.Create(new BaseMetricValue { Time = DateTime.Now, Value = 50 });

            _mock.Verify(repository => repository.Create(It.IsAny<BaseMetricValue>()), Times.AtMostOnce());


            ////arange
            //TimeSpan timeFrom = TimeSpan.FromSeconds(0);
            //TimeSpan timeTo = TimeSpan.FromSeconds(100);

            ////act
            //var result = _controller.GetMetric(timeFrom, timeTo);

            ////assert
            //Assert.IsAssignableFrom<IActionResult>(result);

        }
    }
}
